refactor functions api tests in functionsapitest package
