+++ rest/functionsapitest/graphql_test.go
@@ -1,0 +1,117 @@
+//  Copyright 2022-Present Couchbase, Inc.
+//
+//  Use of this software is governed by the Business Source License included
+//  in the file licenses/BSL-Couchbase.txt.  As of the Change Date specified
+//  in that file, in accordance with the Business Source License, use of this
+//  software will be governed by the Apache License, Version 2.0, included in
+//  the file licenses/APL2.txt.
+
+package functionsapitest
+
+import (
+	"testing"
+
+	"github.com/couchbase/sync_gateway/base"
+	"github.com/couchbase/sync_gateway/db/functions"
+	"github.com/couchbase/sync_gateway/rest"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+const kDummyGraphQLSchema = `
+	type Query {
+		square(n: Int!) : Int!
+	}`
+
+func TestFunctionsConfigGraphQLGetEmpty(t *testing.T) {
+	rt := rest.NewRestTesterForUserQueries(t, rest.DbConfig{})
+	if rt == nil {
+		return
+	}
+	defer rt.Close()
+
+	t.Run("Non-Admin", func(t *testing.T) {
+		response := rt.SendRequest("GET", "/db/_config/graphql", "")
+		assert.Equal(t, 404, response.Result().StatusCode)
+	})
+	t.Run("All", func(t *testing.T) {
+		response := rt.SendAdminRequest("GET", "/db/_config/graphql", "")
+		assert.Equal(t, 404, response.Result().StatusCode)
+	})
+}
+func TestFunctionsConfigGraphQLGet(t *testing.T) {
+	rt := rest.NewRestTesterForUserQueries(t, rest.DbConfig{
+		GraphQL: &functions.GraphQLConfig{
+			Schema:    base.StringPtr(kDummyGraphQLSchema),
+			Resolvers: nil,
+		},
+	})
+	if rt == nil {
+		return
+	}
+	defer rt.Close()
+
+	t.Run("Non-Admin", func(t *testing.T) {
+		response := rt.SendRequest("GET", "/db/_config/graphql", "")
+		assert.Equal(t, 404, response.Result().StatusCode)
+	})
+	t.Run("All", func(t *testing.T) {
+		response := rt.SendAdminRequest("GET", "/db/_config/graphql", "")
+		var body functions.GraphQLConfig
+		require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
+		assert.Equal(t, base.StringPtr(kDummyGraphQLSchema), body.Schema)
+	})
+}
+
+func TestFunctionsConfigGraphQLPut(t *testing.T) {
+	rt := rest.NewRestTesterForUserQueries(t, rest.DbConfig{
+		GraphQL: &functions.GraphQLConfig{
+			Schema:    base.StringPtr(kDummyGraphQLSchema),
+			Resolvers: nil,
+		},
+	})
+	if rt == nil {
+		return
+	}
+	defer rt.Close()
+
+	t.Run("Non-Admin", func(t *testing.T) {
+		response := rt.SendRequest("PUT", "/db/_config/graphql", "{}")
+		assert.Equal(t, 404, response.Result().StatusCode)
+		response = rt.SendRequest("DELETE", "/db/_config/graphql", "{}")
+		assert.Equal(t, 404, response.Result().StatusCode)
+	})
+	t.Run("ReplaceBogus", func(t *testing.T) {
+		response := rt.SendAdminRequest("PUT", "/db/_config/graphql", `{
+			"schema": "obviously not a valid schema ^_^"
+		}`)
+		assert.Equal(t, 400, response.Result().StatusCode)
+	})
+	t.Run("Replace", func(t *testing.T) {
+		response := rt.SendAdminRequest("PUT", "/db/_config/graphql", `{
+			"schema": "type Query {sum(n: Int!) : Int!}",
+			"resolvers": {
+				"Query": {
+					"sum": {
+						"type": "javascript",
+						"code": "function(context,args){return args.n + args.n;}"
+					}
+				}
+			}
+		}`)
+		assert.Equal(t, 200, response.Result().StatusCode)
+
+		response = rt.SendAdminRequest("POST", "/db/_graphql", `{"query": "query{ sum(n:3) }"}`)
+		assert.Equal(t, 200, response.Result().StatusCode)
+		assert.Equal(t, `{"data":{"sum":6}}`, string(response.BodyBytes()))
+	})
+	t.Run("Delete", func(t *testing.T) {
+		response := rt.SendAdminRequest("DELETE", "/db/_config/graphql", "")
+		assert.Equal(t, 200, response.Result().StatusCode)
+
+		assert.Nil(t, rt.GetDatabase().Options.GraphQL)
+
+		response = rt.SendAdminRequest("POST", "/db/_graphql", `{"query": "query{ sum(n:3) }"}`)
+		assert.Equal(t, 503, response.Result().StatusCode)
+	})
+}
