+++ rest/functionsapitest/user_functions_test.go
@@ -1,0 +1,206 @@
+//  Copyright 2022-Present Couchbase, Inc.
+//
+//  Use of this software is governed by the Business Source License included
+//  in the file licenses/BSL-Couchbase.txt.  As of the Change Date specified
+//  in that file, in accordance with the Business Source License, use of this
+//  software will be governed by the Apache License, Version 2.0, included in
+//  the file licenses/APL2.txt.
+
+package functionsapitest
+
+import (
+	"testing"
+
+	"github.com/couchbase/sync_gateway/base"
+	"github.com/couchbase/sync_gateway/db/functions"
+	"github.com/couchbase/sync_gateway/rest"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+// When there's no existing config, API calls return 404:
+func TestFunctionsConfigGetMissing(t *testing.T) {
+	rt := rest.NewRestTesterForUserQueries(t, rest.DbConfig{})
+	if rt == nil {
+		return
+	}
+	defer rt.Close()
+
+	t.Run("Non-Admin", func(t *testing.T) {
+		response := rt.SendRequest("GET", "/db/_config/functions", "")
+		assert.Equal(t, 404, response.Result().StatusCode)
+	})
+	t.Run("All", func(t *testing.T) {
+		response := rt.SendAdminRequest("GET", "/db/_config/functions", "")
+		assert.Equal(t, 404, response.Result().StatusCode)
+	})
+	t.Run("Missing", func(t *testing.T) {
+		response := rt.SendAdminRequest("GET", "/db/_config/functions/cube", "")
+		assert.Equal(t, 404, response.Result().StatusCode)
+	})
+}
+func TestFunctionsConfigGet(t *testing.T) {
+	rt := rest.NewRestTesterForUserQueries(t, rest.DbConfig{
+		UserFunctions: &functions.FunctionsConfig{
+			Definitions: functions.FunctionsDefs{
+				"square": {
+					Type:  "javascript",
+					Code:  "function(context,args){return args.numero * args.numero;}",
+					Args:  []string{"numero"},
+					Allow: &functions.Allow{Channels: []string{"wonderland"}},
+				},
+			},
+		},
+	})
+	if rt == nil {
+		return
+	}
+	defer rt.Close()
+
+	t.Run("Non-Admin", func(t *testing.T) {
+		response := rt.SendRequest("GET", "/db/_config/functions", "")
+		assert.Equal(t, 404, response.Result().StatusCode)
+	})
+	t.Run("All", func(t *testing.T) {
+		response := rt.SendAdminRequest("GET", "/db/_config/functions", "")
+		var body functions.FunctionsConfig
+		require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
+		assert.NotNil(t, body.Definitions["square"])
+	})
+	t.Run("Single", func(t *testing.T) {
+		response := rt.SendAdminRequest("GET", "/db/_config/functions/square", "")
+		var body functions.FunctionConfig
+		require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
+		assert.Equal(t, "function(context,args){return args.numero * args.numero;}", body.Code)
+	})
+	t.Run("Missing", func(t *testing.T) {
+		response := rt.SendAdminRequest("GET", "/db/_config/functions/bogus", "")
+		assert.Equal(t, 404, response.Result().StatusCode)
+	})
+}
+
+func TestFunctionsConfigPut(t *testing.T) {
+	rt := rest.NewRestTesterForUserQueries(t, rest.DbConfig{
+		UserFunctions: &functions.FunctionsConfig{
+			Definitions: functions.FunctionsDefs{
+				"square": {
+					Type:  "javascript",
+					Code:  "function(context,args){return args.numero * args.numero;}",
+					Args:  []string{"numero"},
+					Allow: &functions.Allow{Channels: []string{"wonderland"}},
+				},
+			},
+		},
+	})
+	if rt == nil {
+		return
+	}
+	defer rt.Close()
+
+	t.Run("Non-Admin", func(t *testing.T) {
+		response := rt.SendRequest("PUT", "/db/_config/functions", "{}")
+		assert.Equal(t, 404, response.Result().StatusCode)
+		response = rt.SendRequest("DELETE", "/db/_config/functions", "{}")
+		assert.Equal(t, 404, response.Result().StatusCode)
+	})
+	t.Run("ReplaceAll", func(t *testing.T) {
+		response := rt.SendAdminRequest("PUT", "/db/_config/functions", `{
+			"definitions": {
+				"sum": {"type": "javascript",
+						"code": "function(context,args){return args.numero + args.numero;}",
+						"args": ["numero"],
+						"allow": {"channels": ["*"]}} } }`)
+		assert.Equal(t, 200, response.Result().StatusCode)
+
+		assert.NotNil(t, rt.GetDatabase().Options.UserFunctions.Definitions["sum"])
+		assert.Nil(t, rt.GetDatabase().Options.UserFunctions.Definitions["square"])
+
+		response = rt.SendAdminRequest("GET", "/db/_function/sum?numero=13", "")
+		assert.Equal(t, 200, response.Result().StatusCode)
+		assert.Equal(t, "26", string(response.BodyBytes()))
+
+		response = rt.SendAdminRequest("GET", "/db/_function/square?numero=13", "")
+		assert.Equal(t, 404, response.Result().StatusCode)
+	})
+	t.Run("DeleteAll", func(t *testing.T) {
+		response := rt.SendAdminRequest("DELETE", "/db/_config/functions", "")
+		assert.Equal(t, 200, response.Result().StatusCode)
+
+		assert.Nil(t, rt.GetDatabase().Options.UserFunctions)
+
+		response = rt.SendAdminRequest("GET", "/db/_function/square?numero=13", "")
+		assert.Equal(t, 404, response.Result().StatusCode)
+	})
+}
+
+func TestFunctionsConfigPutOne(t *testing.T) {
+	rt := rest.NewRestTesterForUserQueries(t, rest.DbConfig{
+		UserFunctions: &functions.FunctionsConfig{
+			Definitions: functions.FunctionsDefs{
+				"square": {
+					Type:  "javascript",
+					Code:  "function(context,args){return args.numero * args.numero;}",
+					Args:  []string{"numero"},
+					Allow: &functions.Allow{Channels: []string{"wonderland"}},
+				},
+			},
+		},
+	})
+	if rt == nil {
+		return
+	}
+	defer rt.Close()
+
+	t.Run("Non-Admin", func(t *testing.T) {
+		response := rt.SendRequest("PUT", "/db/_config/functions/square", "{}")
+		assert.Equal(t, 404, response.Result().StatusCode)
+		response = rt.SendRequest("DELETE", "/db/_config/function/square", "{}")
+		assert.Equal(t, 404, response.Result().StatusCode)
+	})
+	t.Run("Bogus", func(t *testing.T) {
+		response := rt.SendAdminRequest("PUT", "/db/_config/functions/square", `[]`)
+		assert.Equal(t, 400, response.Result().StatusCode)
+		response = rt.SendAdminRequest("PUT", "/db/_config/functions/square", `{"ruby": "foo"}`)
+		assert.Equal(t, 400, response.Result().StatusCode)
+	})
+	t.Run("Add", func(t *testing.T) {
+		response := rt.SendAdminRequest("PUT", "/db/_config/functions/sum", `{
+			"type": "javascript",
+			"code": "function(context,args){return args.numero + args.numero;}",
+			"args": ["numero"],
+			"allow": {"channels": ["*"]}
+		}`)
+		assert.Equal(t, 200, response.Result().StatusCode)
+
+		assert.NotNil(t, rt.GetDatabase().Options.UserFunctions.Definitions["sum"])
+		assert.NotNil(t, rt.GetDatabase().Options.UserFunctions.Definitions["square"])
+
+		response = rt.SendAdminRequest("GET", "/db/_function/sum?numero=13", "")
+		assert.Equal(t, "26", string(response.BodyBytes()))
+	})
+	t.Run("ReplaceOne", func(t *testing.T) {
+		response := rt.SendAdminRequest("PUT", "/db/_config/functions/square", `{
+			"type": "javascript",
+			"code": "function(context,args){return -args.n * args.n;}",
+			"args": ["n"],
+			"allow": {"channels": ["*"]}
+		}`)
+		assert.Equal(t, 200, response.Result().StatusCode)
+
+		assert.NotNil(t, rt.GetDatabase().Options.UserFunctions.Definitions["sum"])
+		assert.NotNil(t, rt.GetDatabase().Options.UserFunctions.Definitions["square"])
+
+		response = rt.SendAdminRequest("GET", "/db/_function/square?n=13", "")
+		assert.Equal(t, "-169", string(response.BodyBytes()))
+	})
+	t.Run("DeleteOne", func(t *testing.T) {
+		response := rt.SendAdminRequest("DELETE", "/db/_config/functions/square", "")
+		assert.Equal(t, 200, response.Result().StatusCode)
+
+		assert.Nil(t, rt.GetDatabase().Options.UserFunctions.Definitions["square"])
+		assert.Equal(t, 1, len(rt.GetDatabase().Options.UserFunctions.Definitions))
+
+		response = rt.SendAdminRequest("GET", "/db/_function/square?n=13", "")
+		assert.Equal(t, 404, response.Result().StatusCode)
+	})
+}
